requirements.txt
bokeh==2.4.2
Bottleneck==1.3.5
certifi==2022.9.24
click==8.1.3
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
flit_core==3.6.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
pip==22.3.1
plotly==5.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
setuptools==65.5.0
six==1.16.0
tenacity==8.1.0
tornado==6.2
typing_extensions==4.4.0
wheel==0.37.1
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.11.0
click==8.1.3
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
flit_core==3.6.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
plotly==5.13.0
pyparsing==3.0.9
